Fix network and router
Updating Network in Repository package issued with an
error when trying to get tags. It was due to the invalid object
that was investigated with the proper tag.
Changed old network update logic to read existing network,
check the tag and update object.
Logical Router deletion had bug with invalid type as well.

Change-Id: Ib6ca82a3c85955ee1bcf8bb4469bb011979c00b1