Ignoring tagged ports
New option in configuration, `management_port_tags`, defines the names
of the tags on ports for plugin to ignore.

This allows to ommit some plugins connected to switch that have
different role than data plane, like management. Those ports are not
longer configured by the plugin and are safe to have own configuration,
even if this configuration was created directly on switch, not by
Contrail.

Change-Id: I685880ce4f9b5fe805f813543c324efc4dd98eec